+++ src/api/objects/trending_with_history.ts
@@ -61,8 +61,8 @@ export function uniquifyTrendingObjs<T extends TrendingWithHistory>(
         const key = uniqueKey(obj).toLowerCase();
 
         if (unique[key]) {
-            unique[key].numToots = Math.max(unique[key].numToots || 0, obj.numToots || 0);
             unique[key].numAccounts = Math.max(unique[key].numAccounts || 0, obj.numAccounts || 0);
+            unique[key].numToots = Math.max(unique[key].numToots || 0, obj.numToots || 0);
         } else {
             unique[key] = obj;
         }
@@ -70,6 +70,7 @@ export function uniquifyTrendingObjs<T extends TrendingWithHistory>(
         return unique;
     }, {} as Record<string, TrendingWithHistory>);
 
+    // TODO: should this sort by numToots too instead? Usually we sort things by numToots but here we want to sort by numAccounts
     const sortedObjs = Object.values(urlObjs).sort((a, b) => (b.numAccounts || 0) - (a.numAccounts || 0));
     return sortedObjs as T[];
 };