+++ src/helpers/math_helper.ts
@@ -1,7 +1,7 @@
 /*
  * Math and numbers.
  */
-import { byteString, NUMBER_REGEX } from "./string_helpers";
+import { NUMBER_REGEX, byteString, isNull } from "./string_helpers";
 import { strBytes } from "./log_helpers";
 import { sumArray } from "./collection_helpers";
 import { type StringNumberDict } from "../types";
@@ -69,7 +69,7 @@ export function sizeFromBufferByteLength(obj: object): number {
 
 // Not 100% accurate. From https://gist.github.com/rajinwonderland/36887887b8a8f12063f1d672e318e12e
 export function sizeOf(obj: any, sizes: BytesDict): number {
-    if (obj === null || obj === undefined) return 0;
+    if (isNull(obj)) return 0;
     let bytes = 0;
 
     switch (typeof obj) {