+++ src/helpers/string_helpers.ts
@@ -46,8 +46,10 @@ export const MEDIA_TYPES: mastodon.v1.MediaAttachmentType[] = [
 
 // Check if it's a string
 export const isString = (s: unknown) => typeof s === 'string'; // || s instanceof String; // TODO: wtf is String about?
+// Check if null or undefined
+export const isNull = (s: any) => s === null || s === undefined;
 // Check if it's empty (all whitespace or null or undefined)
-export const isEmptyStr = (s: string | null | undefined) => s === null || s === undefined || s.trim() === '';
+export const isEmptyStr = (s: string | null | undefined) => isNull(s) || s!.trim() === '';
 
 // Alphabetize an array of strings
 export const alphabetize = (arr: string[]) => arr.sort(compareStr);
@@ -63,7 +65,7 @@ export const bracketed = (str: string): string => str.startsWith('[') ? str : `[
 // Prefix a string with [Brackets] and a space
 export const prefixed = (prefix: string, msg: string) => `${bracketed(prefix)} ${msg}`;
 // Doublequotes
-export const quoted = (str: string | null): string => str == null ? NULL : `"${str}"`;
+export const quoted = (str: string | null): string => isNull(str) ? NULL : `"${str}"`;
 
 // Collapse whitespace in a string
 export const collapseWhitespace = (str: string) => str.replace(WHITESPACE_REGEX, " ").replace(/\s,/g,  ",").trim();