+++ src/types.ts
@@ -8,6 +8,7 @@ import Account from './api/objects/account';
 import BooleanFilter, { BooleanFilterArgs, BooleanFilterName } from './filters/boolean_filter';
 import NumericFilter, { NumericFilterArgs } from './filters/numeric_filter';
 import Scorer from './scorer/scorer';
+import TagList from './api/tag_list';
 import Toot, { SerializableToot } from './api/objects/toot';
 import { CacheKey, NonScoreWeightName, ScoreName, TagTootsCacheKey } from './enums';
 
@@ -165,7 +166,7 @@ export interface TrendingLink extends mastodon.v1.TrendLink, TrendingHistory {};
 export type TrendingData = {
     links: TrendingLink[];
     servers: MastodonInstances;
-    tags: TagWithUsageCounts[];
+    tags: TagList;
     toots: Toot[];
 };